Backs out networking changes. Adds Ansible provisioning for rebooting the VM from inside of the VM
It turns out restarting networking was just removing the interfaces and
traffic was only working through the NAT interface.  The only way to fix
the issue with UCS is to reboot the box from inside the VM.  To do this
provisioning via Ansible is added that reboots the VM and then waits for
it to come back.  It then has to remount the vagrant synced folder.