Adds NAT support from Admin to Public network
Detects where the default gateway is.  If it is on the Public network it
NATs to the Admin Network (eth0) to Public (eth3). This is necessary for
PXE booting nodes with only eth0 access to get out to the internet.
If Admin Network already has internet access, behavior is still the same
as previous behavior.