Includes fixes for DNS and puppet on nodes
Now DNS is correctly set (should be the foreman VM) and puppet is
installed correctly with this patch.  It looks like things work after
this and the node will report to foreman that it built.  May want to
consider removing yum update all, and adding puppet kickoff at the end
of the provision.