Fixes the problem with DNS not resolving to foreman VM and adds puppet startup
Looks like the issue with DNS was because the VM has the extra NAT'ed
interface which was getting a NAT DNS proxy which doesn't work with
looking up the hostname for "foreman-server.opnfv.com".  Added
PEERDNS=no to that interface to stop that problem.

Also added puppet start to the end of the provisioning.  Traditional
baremetal machines are rebooted at the end of the provision.  I would
like to avoid reboot if possible so seeing if we can just start
puppet.service to kickstart the agent.