Adds private ip to /etc/hosts
Noticed an issue where the short hostname "oscontroller1" was being used
by rabbitmq as a lookup to connect to.  Needed to add the short name
into /etc/hosts to get it to start rabbitmq-server.  Probably an issue
with the dns server, but this fix makes us rely slightly less on that
server.